modified to work with new version of Split 2
The RunCam Split 2 case changed to have a taller case with a non-flat
shoulder. That new case doesn't fit in the cradle I made for the
original case. So, I depended that pocket to have 0.1" clearance away
from the shoulder of the new case. The pocket is now just insurance
against a screw failure. The #0 screws are the things actually keeping
the cameras in place.